backend/requirements.txt
fastapi[standard]
python-multipart
email-validator
passlib[bcrypt]
tenacity
pydantic
emails
jinja2
alembic
httpx
psycopg
sqlmodel
bcrypt
pydantic-settings
sentry-sdk[fastapi]
pyjwt
pytest
mypy
ruff
pre-commit
types-passlib
coverage
psycopg[binary]
fastapi_mail